+++ src/app/api/apiServices/httpRequest.tsx
@@ -4,7 +4,7 @@ import axios from 'axios';
 
 // Custom fetch GET method using axios
 export const get = async (apiPath: string, option = {}) => {
-  apiPath = process.env.NEXT_PUBLIC_BASE_URL + apiPath;
+  apiPath = process.env.NEXT_PUBLIC_BASE_URL + '/' + apiPath;
 
   const res = await axios.get(apiPath, option);
   return res.data;
