+++ src/components/post/PostContent.module.scss
@@ -31,3 +31,55 @@
   white-space: nowrap;
   cursor: pointer;
 }
+
+.spoiler-overlay {
+  position: absolute;
+  width: 100%;
+  height: 75%;
+  top: 25%;
+  left: 0;
+  border-radius: 10px;
+  -webkit-transition: all 0.1s ease;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition: all 0.1s ease;
+  pointer-events: none;
+  background-color: rgba(0, 0, 0, 0.4);
+  backdrop-filter: blur(44px);
+  cursor: pointer;
+}
+
+.spoiler-button {
+  position: absolute;
+  left: 50%;
+  top: 60%;
+  -webkit-transform: translate(-50%, -50%);
+  transform: translate(-50%, -50%);
+  text-transform: uppercase;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+  -ms-flex-direction: column;
+  flex-direction: column;
+  -webkit-box-align: center;
+  -ms-flex-align: center;
+  align-items: center;
+  font-weight: 600;
+  cursor: pointer;
+  z-index: 1;
+  padding: 8px 12px;
+  border-radius: 20px;
+  letter-spacing: 0.5px;
+  font-size: 15px;
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+  background-color: #4e5d78;
+  &:hover {
+    background-color: #324261;
+  }
+}
