+++ src/components/post/Video.tsx
@@ -1,20 +1,13 @@
-'use client'
+'use client';
 
-import { useEffect, useState } from "react";
+import { useEffect, useState } from 'react';
 
-export function VideoComponent({ videoUrl }: { videoUrl: string }) {
+export function VideoComponent({ videoUrl, className }: { videoUrl: string; className: string }) {
   const [hydrated, setHydrated] = useState(false);
 
   useEffect(() => {
     setHydrated(true);
   }, []);
 
-  return (
-    <div>
-      {hydrated && (
-        <video src={videoUrl} controls className="rounded-2xl" />
-      )}
-    </div>
-  );
+  return <div>{hydrated && <video src={videoUrl} controls className={`rounded-2xl ${className}`} />}</div>;
 }
-
